+++ apps/cowswap-frontend/src/modules/tradeFlow/hooks/useHandleSwap.ts
@@ -15,7 +15,6 @@ import {
   TradeWidgetActions,
   logTradeFlow,
   useDerivedTradeState,
-  useSolanaWrapAndDelegateDoTrade,
   useTradeFlowAnalytics,
   useTradePriceImpact,
 } from 'modules/trade'
@@ -117,12 +116,6 @@ export function useHandleSwap(
     onUserInput,
     derivedTradeState?.slippage,
   ])
-
-  const solanaWrapAndDelegateDoTrade = useSolanaWrapAndDelegateDoTrade()
-
-  if (solanaWrapAndDelegateDoTrade) {
-    return solanaWrapAndDelegateDoTrade
-  }
 
   return { callback, contextIsReady }
 }